Add harvester cli runner
